pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lovstudio-skill-helper"
version = "0.6.0"
description = "Lovstudio skill helper — activate and run paid Lovstudio skills locally. Decryption keys never touch disk."
requires-python = ">=3.9"
dependencies = [
  "cryptography>=42",
  "pyyaml>=6",
]

[project.scripts]
lovstudio-skill-helper = "lovstudio_skill_helper.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/lovstudio_skill_helper"]
